refactor(recording): inline platform recorder implementations
The suffix files now own both platform resolution and implementation. This removes the extra recorder implementation files and leaves ManualRecorderLive as the single public manual recorder surface.
